+++ changelog.md
@@ -1,11 +1,16 @@
 # changelog
 
+## 0.80.1
+
+- add build config option to mock SvelteKit paths in production
+  ([#374](https://github.com/feltjs/gro/pull/374))
+
 ## 0.80.0
 
 - **break**: replace `cheap-watch` with `chokidar` and `fast-glob`
   ([#386](https://github.com/feltjs/gro/pull/386))
-- add build config option to mock SvelteKit paths in production
-  ([#374](https://github.com/feltjs/gro/pull/374))
+- fix `gen` error on startup for projects with no gen modules
+  ([#387](https://github.com/feltjs/gro/pull/387))
 
 ## 0.79.1